+++ nflows/transforms/UMNN/__init__.py
@@ -1,1 +1,3 @@
-from nflows.transforms.UMNN.MonotonicNormalizer import MonotonicNormalizer, IntegrandNet
+from nflows.transforms.UMNN.monotonicnormalizer import MonotonicNormalizer, IntegrandNet
+from nflows.transforms.UMNN.autoregressive import MaskedUMNNAutoregressiveTransform
+from nflows.transforms.UMNN.coupling import UMNNCouplingTransform
